Updated to use a base class and cleaned up tests
