+++ python/idsse_common/idsse/common/aws_utils.py
@@ -2,10 +2,12 @@
 # -------------------------------------------------------------------------------
 # Created on Tue Feb 14 2023
 #
-# Copyright (c) 2023 Regents of the University of Colorado. All rights reserved.
+# Copyright (c) 2023 Colorado State University. All rights reserved. (1)
+# Copyright (c) 2023 Regents of the University of Colorado. All rights reserved. (2)
 #
 # Contributors:
-#     Geary J Layne
+#     Geary Layne (2)
+#     Paul Hamer (1)
 #
 # -------------------------------------------------------------------------------
 
@@ -17,11 +19,12 @@ from datetime import datetime, timedelta, UTC
 
 from .path_builder import PathBuilder
 from .utils import TimeDelta, datetime_gen, exec_cmd
+from .protocol_utils import ProtocolUtils
 
 logger = logging.getLogger(__name__)
 
 
-class AwsUtils():
+class AwsUtils(ProtocolUtils):
     """AWS Utility Class"""
 
     def __init__(self,
@@ -29,7 +32,8 @@ class AwsUtils():
                  subdir: str,
                  file_base: str,
                  file_ext: str) -> None:
-        self.path_builder = PathBuilder(basedir, subdir, file_base, file_ext)
+        super().__init__(basedir, subdir, file_base, file_ext)
+
 
     def get_path(self, issue: datetime, valid: datetime) -> str:
         """Delegates to instant PathBuilder to get full path given issue and valid
@@ -41,17 +45,30 @@ class AwsUtils():
         Returns:
             str: Absolute path to file or object
         """
-        lead = TimeDelta(valid-issue)
-        return self.path_builder.build_path(issue=issue, valid=valid, lead=lead)
+        return super().get_path(issue, valid)
 
-    def aws_ls(self, path: str, prepend_path: bool = True) -> Sequence[str]:
-        """Execute an 'ls' on the AWS s3 bucket specified by path
+    def ls(self, path: str, prepend_path: bool = True) -> Sequence[str]:
+        """Execute a 'ls' on the AWS s3 bucket specified by path
 
         Args:
             path (str): s3 bucket
+            prepend_path (bool): Add to the filename
 
         Returns:
-            Sequence[str]: The results sent to stdout from executing an 'ls' on passed path
+            Sequence[str]: The results sent to stdout from executing a 'ls' on passed path
+        """
+        return self.aws_ls(path, prepend_path)
+
+    @staticmethod
+    def aws_ls(path: str, prepend_path: bool = True) -> Sequence[str]:
+        """Execute a 'ls' on the AWS s3 bucket specified by path
+
+        Args:
+            path (str): s3 bucket
+            prepend_path (bool): Add to the filename
+
+        Returns:
+            Sequence[str]: The results sent to stdout from executing a 'ls' on passed path
         """
         try:
             commands = ['s5cmd',  '--no-sign-request', 'ls', path]
@@ -65,12 +82,12 @@ class AwsUtils():
             return [os.path.join(path, filename.split(' ')[-1]) for filename in commands_result]
         return [filename.split(' ')[-1] for filename in commands_result]
 
-    def aws_cp(self,
-               path: str,
+    @staticmethod
+    def aws_cp(path: str,
                dest: str,
                concurrency: int | None = None,
                chunk_size: int | None = None) -> bool:
-        """Execute an 'cp' on the AWS s3 bucket specified by path, dest. Attempts to use
+        """Execute a 'cp' on the AWS s3 bucket specified by path, dest. Attempts to use
         [s5cmd](https://github.com/peak/s5cmd) to copy the file from S3 with parallelization,
         but falls back to (slower) aws-cli if s5cmd is not installed or throws an error.
 
@@ -78,7 +95,7 @@ class AwsUtils():
             path (str): Relative or Absolute path to the object to be copied
             dest (str): The destination location
             concurrency (optional, int): Number of parallel threads for s5cmd to use to copy
-                the file down from AWS (may be helpful to tweak for large files).
+                the file down from AWS (maybe helpful to tweak for large files).
                 Default is None (s5cmd default).
             chunk_size (optional, int): Size of chunks (in MB) for s5cmd to split up the source AWS
                 S3 file so it can download quicker with more threads.
@@ -112,6 +129,7 @@ class AwsUtils():
             finally:
                 pass
 
+
     def check_for(self, issue: datetime, valid: datetime) -> tuple[datetime, str] | None:
         """Checks if an object passed issue/valid exists
 
@@ -121,20 +139,10 @@ class AwsUtils():
 
         Returns:
             [tuple[datetime, str] | None]: A tuple of the valid date/time (indicated by object's
-                                            location) and location (path) of a object, or None
+                                            location) and location (path) of an object, or None
                                             if object does not exist
         """
-        lead = TimeDelta(valid-issue)
-        file_path = self.get_path(issue, valid)
-        dir_path = os.path.dirname(file_path)
-        filenames = self.aws_ls(file_path, prepend_path=False)
-        filename = self.path_builder.build_filename(issue=issue, valid=valid, lead=lead)
-        for fname in filenames:
-            # Support wildcard matches - used for '?' as a single wildcard character in
-            # issue/valid time specs.
-            if fnmatch.fnmatch(os.path.basename(fname), filename):
-                return (valid, os.path.join(dir_path, fname))
-        return None
+        return super().check_for(issue, valid)
 
     def get_issues(self,
                    num_issues: int = 1,
@@ -153,32 +161,8 @@ class AwsUtils():
         Returns:
             Sequence[datetime]: A sequence of issue date/times
         """
-        zero_time_delta = timedelta(seconds=0)
-        if time_delta == zero_time_delta:
-            raise ValueError('Time delta must be non zero')
+        return super().get_issues(num_issues, issue_start, issue_end, time_delta)
 
-        issues_set: set[datetime] = set()
-        if issue_start:
-            datetimes = datetime_gen(issue_end, time_delta, issue_start, num_issues)
-        else:
-            # check if time delta is positive, if so make negative
-            if time_delta > zero_time_delta:
-                time_delta = timedelta(seconds=-1.0 * time_delta.total_seconds())
-            datetimes = datetime_gen(issue_end, time_delta)
-        for issue_dt in datetimes:
-            if issue_start and issue_dt < issue_start:
-                break
-            try:
-                dir_path = self.path_builder.build_dir(issue=issue_dt)
-                issues = {self.path_builder.get_issue(file_path)
-                          for file_path in self.aws_ls(dir_path)
-                          if file_path.endswith(self.path_builder.file_ext)}
-                issues_set.update(issues)
-                if num_issues and len(issues_set) >= num_issues:
-                    break
-            except PermissionError:
-                pass
-        return sorted(issues_set, reverse=True)[:num_issues]
 
     def get_valids(self,
                    issue: datetime,
@@ -198,27 +182,4 @@ class AwsUtils():
                                             object's location) and the object's location (path).
                                             Empty Sequence if no valids found for given time range.
         """
-        if valid_start and valid_start == valid_end:
-            valids_and_filenames = self.check_for(issue, valid_start)
-            return [valids_and_filenames] if valids_and_filenames is not None else []
-
-        dir_path = self.path_builder.build_dir(issue=issue)
-        valid_and_file = [(self.path_builder.get_valid(file_path), file_path)
-                          for file_path in self.aws_ls(dir_path)
-                          if file_path.endswith(self.path_builder.file_ext)]
-
-        if valid_start:
-            if valid_end:
-                valid_and_file = [(valid, filename)
-                                  for valid, filename in valid_and_file
-                                  if valid_start <= valid <= valid_end]
-            else:
-                valid_and_file = [(valid, filename)
-                                  for valid, filename in valid_and_file
-                                  if valid >= valid_start]
-        elif valid_end:
-            valid_and_file = [(valid, filename)
-                              for valid, filename in valid_and_file
-                              if valid <= valid_end]
-
-        return valid_and_file
+        return super().get_valids(issue, valid_start, valid_end)
