+++ python/idsse_common/test/test_aws_utils.py
@@ -8,6 +8,7 @@
 # Contributors:
 #     Mackenzie Grimes (1)
 #     Geary Layne (2)
+#     Paul Hamer (1)
 #
 # ----------------------------------------------------------------------------------
 # pylint: disable=missing-function-docstring,redefined-outer-name,pointless-statement
@@ -56,7 +57,10 @@ def aws_utils_with_wild() -> AwsUtils:
 @fixture
 def mock_exec_cmd(monkeypatch: MonkeyPatch) -> Mock:
     def get_files_for_dir(args: Iterable[str]) -> Sequence[str]:
-        hour = args[-1].split('/')[-3]
+        if args[-1].endswith('grib2') or args[-1].endswith('/'):
+            hour = args[-1].split('/')[-3]
+        else:
+            hour = args[-1].split('/')[-2]
         return [f'blend.t{hour}z.core.f002.co.grib2',
                 f'blend.t{hour}z.core.f003.co.grib2',
                 f'blend.t{hour}z.core.f004.co.grib2']
@@ -155,8 +159,8 @@ def test_get_issues(aws_utils: AwsUtils, mock_exec_cmd):
     result = aws_utils.get_issues(
         issue_start=EXAMPLE_ISSUE, issue_end=EXAMPLE_VALID, num_issues=2)
     assert len(result) == 2
-    assert result[0] == EXAMPLE_VALID
-    assert result[1] == EXAMPLE_VALID - timedelta(hours=1)
+    assert result[0] == EXAMPLE_VALID - timedelta(hours=1)
+    assert result[1] == EXAMPLE_VALID
 
 
 def test_get_issues_with_same_start_stop(aws_utils: AwsUtils, mock_exec_cmd):
